# main.py
from enum import Enum
from typing import List
import random

from fastapi import FastAPI, Query, Security, Depends
from fastapi.security import APIKeyHeader
import uvicorn

app = FastAPI()
api_key_header = APIKeyHeader(name="X-Api-Key", auto_error=True)


class Animal(str, Enum):
    bird = "bird"
    cat = "cat"
    fish = "fish"


async def get_user(api_key=Security(api_key_header)):
    if api_key != "abc123":
        raise HTTPException(
            status_code=HTTP_403_FORBIDDEN
        )
    return "Bob"


@app.get("/generate/{animal}", response_model=List[str])
async def generate_thing(
    animal: Animal,
    adjectives: int = Query(1, ge=0, le=3),
    count: int = Query(1, ge=1, le=5),
    _user=Depends(get_user),
):
    return [generate_animal(animal, adjectives) for _x in range(count)]


def generate_animal(_animal: Animal, adjectives: int):
    # TODO generate actual different animal types
    return f"{' '.join([random.choice(adjective_words) for _x in range(adjectives)] + [random.choice(animal_words)])}"


adjective_words = [
    "Able",
    "Adept",
    "Aged",
    "Agile",
    "Ajar",
    "Alert",
    "Alive",
    "All",
    "Ample",
    "Angry",
    "Any",
    "Apt",
    "Arid",
    "Aware",
    "Awful",
    "Bad",
    "Back",
    "Baggy",
    "Bare",
    "Basic",
    "Best",
    "Big",
    "Black",
    "Bland",
    "Blank",
    "Bleak",
    "Blind",
    "Blond",
    "Blue",
    "Bogus",
    "Bold",
    "Bony",
    "Bossy",
    "Both",
    "Bowed",
    "Brave",
    "Brief",
    "Brisk",
    "Brown",
    "Bulky",
    "Bumpy",
    "Burly",
    "Busy",
    "Calm",
    "Cheap",
    "Chief",
    "Clean",
    "Clear",
    "Close",
    "Cold",
    "Cool",
    "Corny",
    "Crazy",
    "Crisp",
    "Cruel",
    "Curly",
    "Curvy",
    "Cute",
    "Damp",
    "Dark",
    "Dead",
    "Dear",
    "Deep",
    "Dense",
    "Dim",
    "Dirty",
    "Dizzy",
    "Dopey",
    "Drab",
    "Dry",
    "Dual",
    "Dull",
    "Each",
    "Eager",
    "Early",
    "Easy",
    "Empty",
    "Equal",
    "Even",
    "Every",
    "Evil",
    "Faint",
    "Fair",
    "Fake",
    "False",
    "Fancy",
    "Far",
    "Fast",
    "Fat",
    "Fatal",
    "Few",
    "Fine",
    "Firm",
    "First",
    "Fixed",
    "Flaky",
    "Flat",
    "Fluid",
    "Fond",
    "Frail",
    "Frank",
    "Free",
    "Fresh",
    "Front",
    "Full",
    "Funny",
    "Fussy",
    "Fuzzy",
    "Giant",
    "Giddy",
    "Glass",
    "Glum",
    "Good",
    "Grand",
    "Grave",
    "Gray",
    "Great",
    "Green",
    "Grim",
    "Grimy",
    "Gross",
    "Grown",
    "Gummy",
    "Hairy",
    "Half",
    "Handy",
    "Happy",
    "Hard",
    "Harsh",
    "Hasty",
    "Heavy",
    "Hefty",
    "High",
    "Hot",
    "Huge",
    "Husky",
    "Icky",
    "Icy",
    "Ideal",
    "Idle",
    "Ill",
    "Itchy",
    "Jaded",
    "Joint",
    "Jolly",
    "Juicy",
    "Jumbo",
    "Jumpy",
    "Keen",
    "Key",
    "Kind",
    "Known",
    "Kooky",
    "Lame",
    "Lanky",
    "Large",
    "Last",
    "Late",
    "Lazy",
    "Lean",
    "Leafy",
    "Left",
    "Legal",
    "Light",
    "Limp",
    "Lined",
    "Live",
    "Livid",
    "Lone",
    "Long",
    "Loose",
    "Lost",
    "Loud",
    "Low",
    "Loyal",
    "Lucky",
    "Lumpy",
    "Mad",
    "Major",
    "Male",
    "Mealy",
    "Mean",
    "Meaty",
    "Meek",
    "Merry",
    "Messy",
    "Mild",
    "Milky",
    "Minor",
    "Minty",
    "Misty",
    "Mixed",
    "Moist",
    "Moral",
    "Muddy",
    "Murky",
    "Mushy",
    "Musty",
    "Muted",
    "Naive",
    "Nasty",
    "Near",
    "Neat",
    "Needy",
    "New",
    "Next",
    "Nice",
    "Nifty",
    "Nippy",
    "Noisy",
    "Noted",
    "Novel",
    "Numb",
    "Nutty",
    "Obese",
    "Oily",
    "Odd",
    "Old",
    "Only",
    "Open",
    "Other",
    "Our",
    "Oval",
    "Pale",
    "Past",
    "Perky",
    "Pesky",
    "Petty",
    "Phony",
    "Pink",
    "Plain",
    "Plump",
    "Plush",
    "Poor",
    "Posh",
    "Prime",
    "Prize",
    "Proud",
    "Puny",
    "Pure",
    "Pushy",
    "Quick",
    "Quiet",
    "Rapid",
    "Rare",
    "Rash",
    "Raw",
    "Ready",
    "Real",
    "Red",
    "Regal",
    "Rich",
    "Rigid",
    "Right",
    "Ripe",
    "Rosy",
    "Rough",
    "Round",
    "Rowdy",
    "Royal",
    "Ruddy",
    "Rude",
    "Runny",
    "Rural",
    "Rusty",
    "Sad",
    "Safe",
    "Salty",
    "Same",
    "Sandy",
    "Sane",
    "Scaly",
    "Scary",
    "Shady",
    "Sharp",
    "Shiny",
    "Short",
    "Showy",
    "Shy",
    "Sick",
    "Silky",
    "Silly",
    "Slim",
    "Slimy",
    "Slow",
    "Small",
    "Smart",
    "Smug",
    "Soft",
    "Soggy",
    "Solid",
    "Some",
    "Sore",
    "Soupy",
    "Sour",
    "Spicy",
    "Spry",
    "Staid",
    "Stale",
    "Stark",
    "Steep",
    "Stiff",
    "Steel",
    "Sunny",
    "Super",
    "Sweet",
    "Swift",
    "Tall",
    "Tame",
    "Tan",
    "Tart",
    "Tasty",
    "Taut",
    "Tense",
    "Tepid",
    "Testy",
    "That",
    "These",
    "Thick",
    "Thin",
    "Third",
    "This",
    "Those",
    "Tidy",
    "Tight",
    "Tiny",
    "Tired",
    "Torn",
    "Total",
    "Tough",
    "Trim",
    "True",
    "Tubby",
    "Twin",
    "Ugly",
    "Unfit",
    "Upset",
    "Urban",
    "Used",
    "Utter",
    "Vague",
    "Vain",
    "Valid",
    "Vapid",
    "Vast",
    "Vital",
    "Vivid",
    "Wan",
    "Warm",
    "Wary",
    "Wavy",
    "Weak",
    "Weary",
    "Wee",
    "Weepy",
    "Weird",
    "Wet",
    "Which",
    "White",
    "Whole",
    "Wide",
    "Wild",
    "Windy",
    "Wiry",
    "Wise",
    "Witty",
    "Woozy",
    "Wordy",
    "Worn",
    "Worse",
    "Worst",
    "Wrong",
    "Wry",
    "Young",
    "Yummy",
    "Zany",
    "Zesty",
]

# noinspection SpellCheckingInspection
animal_words = [
    "Aardvark",
    "Abyssinian",
    "Affenpinscher",
    "Akbash",
    "Akita",
    "Albatross",
    "Alligator",
    "Angelfish",
    "Ant",
    "Anteater",
    "Antelope",
    "Armadillo",
    "Avocet",
    "Axolotl",
    "Baboon",
    "Badger",
    "Balinese",
    "Bandicoot",
    "Barb",
    "Barnacle",
    "Barracuda",
    "Bat",
    "Beagle",
    "Bear",
    "Beaver",
    "Beetle",
    "Binturong",
    "Bird",
    "Birman",
    "Bison",
    "Bloodhound",
    "Bobcat",
    "Bombay",
    "Bongo",
    "Bonobo",
    "Booby",
    "Budgerigar",
    "Buffalo",
    "Bulldog",
    "Bullfrog",
    "Burmese",
    "Butterfly",
    "Caiman",
    "Camel",
    "Capybara",
    "Caracal",
    "Cassowary",
    "Cat",
    "Caterpillar",
    "Catfish",
    "Centipede",
    "Chameleon",
    "Chamois",
    "Cheetah",
    "Chicken",
    "Chihuahua",
    "Chimpanzee",
    "Chinchilla",
    "Chinook",
    "Chipmunk",
    "Cichlid",
    "Coati",
    "Cockroach",
    "Collie",
    "Coral",
    "Cougar",
    "Cow",
    "Coyote",
    "Crab",
    "Crane",
    "Crocodile",
    "Cuscus",
    "Cuttlefish",
    "Dachshund",
    "Dalmatian",
    "Deer",
    "Dhole",
    "Dingo",
    "Discus",
    "Dodo",
    "Dog",
    "Dolphin",
    "Donkey",
    "Dormouse",
    "Dragonfly",
    "Drever",
    "Duck",
    "Dugong",
    "Dunker",
    "Eagle",
    "Earwig",
    "Echidna",
    "Elephant",
    "Emu",
    "Falcon",
    "Ferret",
    "Fish",
    "Flamingo",
    "Flounder",
    "Fly",
    "Fossa",
    "Fox",
    "Frigatebird",
    "Frog",
    "Gar",
    "Gecko",
    "Gerbil",
    "Gharial",
    "Gibbon",
    "Giraffe",
    "Goat",
    "Goose",
    "Gopher",
    "Gorilla",
    "Grasshopper",
    "Greyhound",
    "Grouse",
    "Guppy",
    "Hamster",
    "Hare",
    "Harrier",
    "Havanese",
    "Hedgehog",
    "Heron",
    "Himalayan",
    "Hippopotamus",
    "Horse",
    "Human",
    "Hummingbird",
    "Hyena",
    "Ibis",
    "Iguana",
    "Impala",
    "Indri",
    "Insect",
    "Jackal",
    "Jaguar",
    "Javanese",
    "Jellyfish",
    "Kakapo",
    "Kangaroo",
    "Kingfisher",
    "Kiwi",
    "Koala",
    "Kudu",
    "Labradoodle",
    "Ladybird",
    "Lemming",
    "Lemur",
    "Leopard",
    "Liger",
    "Lion",
    "Lionfish",
    "Lizard",
    "Llama",
    "Lobster",
    "Lynx",
    "Macaw",
    "Magpie",
    "Maltese",
    "Manatee",
    "Mandrill",
    "Markhor",
    "Mastiff",
    "Mayfly",
    "Meerkat",
    "Millipede",
    "Mole",
    "Molly",
    "Mongoose",
    "Mongrel",
    "Monkey",
    "Moorhen",
    "Moose",
    "Moth",
    "Mouse",
    "Mule",
    "Neanderthal",
    "Newfoundland",
    "Newt",
    "Nightingale",
    "Numbat",
    "Ocelot",
    "Octopus",
    "Okapi",
    "Olm",
    "Opossum",
    "Orangutan",
    "Ostrich",
    "Otter",
    "Oyster",
    "Pademelon",
    "Panther",
    "Parrot",
    "Peacock",
    "Pekingese",
    "Pelican",
    "Penguin",
    "Persian",
    "Pheasant",
    "Pig",
    "Pika",
    "Pike",
    "Piranha",
    "Platypus",
    "Pointer",
    "Poodle",
    "Porcupine",
    "Possum",
    "Prawn",
    "Puffin",
    "Pug",
    "Puma",
    "Quail",
    "Quetzal",
    "Quokka",
    "Quoll",
    "Rabbit",
    "Raccoon",
    "Ragdoll",
    "Rat",
    "Rattlesnake",
    "Reindeer",
    "Rhinoceros",
    "Robin",
    "Rottweiler",
    "Salamander",
    "Saola",
    "Scorpion",
    "Seahorse",
    "Seal",
    "Serval",
    "Sheep",
    "Shrimp",
    "Siamese",
    "Siberian",
    "Skunk",
    "Sloth",
    "Snail",
    "Snake",
    "Snowshoe",
    "Somali",
    "Sparrow",
    "Sponge",
    "Squid",
    "Squirrel",
    "Starfish",
    "Stingray",
    "Stoat",
    "Swan",
    "Tang",
    "Tapir",
    "Tarsier",
    "Termite",
    "Tetra",
    "Tiffany",
    "Tiger",
    "Tortoise",
    "Toucan",
    "Tropicbird",
    "Tuatara",
    "Turkey",
    "Uakari",
    "Uguisu",
    "Umbrellabird",
    "Vulture",
    "Wallaby",
    "Walrus",
    "Warthog",
    "Wasp",
    "Weasel",
    "Whippet",
    "Wildebeest",
    "Wolf",
    "Wolverine",
    "Wombat",
    "Woodlouse",
    "Woodpecker",
    "Wrasse",
    "Yak",
    "Zebra",
    "Zebu",
    "Zonkey",
    "Zorse",
]

if __name__ == "__main__":
    # Start development server
    app.debug = True
    uvicorn.run(
        app,
        debug=True,
    )